chore: make transaction tools directly executable
